fix(maps): update imports to renamed Legend components
Update ChoroplethMap and SymbolMap to import from renamed Legend
components (LegendOrdinal and LegendScale).
