+++ lib/maps/SymbolMap/SymbolMap.vue
@@ -18,7 +18,7 @@ import {
 } from 'lodash'
 
 import config from '@/config'
-import OrdinalLegend from '@/components/Legend/OrdinalLegend.vue'
+import OrdinalLegend from '@/components/Legend/LegendOrdinal.vue'
 import { getChartProps, useChart } from '@/composables/useChart'
 import {
   ComponentPublicInstance,
